Ml-SAAS--main/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
